[DRUP-560] Abstract the Prepaid balance controller.
