+++ modules/apigee_m10n_teams/apigee_m10n_teams.routing.yml
@@ -1,7 +1,7 @@
 apigee_monetization_teams.billing:
   path: /teams/{team}/monetization/billing
   defaults:
-    _controller: \Drupal\apigee_m10n_teams\Controller\TeamPrepaidBalanceController::render
+    _controller: \Drupal\apigee_m10n_teams\Controller\TeamPrepaidBalanceController::teamBalancePage
     _title: 'Prepaid balance'
   requirements:
     _permission: 'view mint prepaid reports'
