+++ modules/apigee_m10n_teams/src/Controller/TeamPrepaidBalanceController.php
@@ -19,11 +19,7 @@
 
 namespace Drupal\apigee_m10n_teams\Controller;
 
-use Apigee\Edge\Api\Monetization\Entity\PrepaidBalanceInterface;
 use Drupal\apigee_edge_teams\Entity\TeamInterface;
-use Drupal\apigee_m10n\Controller\PrepaidBalanceController;
-use Drupal\apigee_m10n\Form\PrepaidBalanceRefreshForm;
-use Drupal\apigee_m10n_teams\TeamSdkControllerFactoryAwareTrait;
 
 /**
  * Controller for team balances.
@@ -31,16 +27,7 @@ use Drupal\apigee_m10n_teams\TeamSdkControllerFactoryAwareTrait;
  * This is modeled after an entity list builder with some additions.
  * See: `\Drupal\Core\Entity\EntityListBuilder`
  */
-class TeamPrepaidBalanceController extends PrepaidBalanceController {
-
-  use TeamSdkControllerFactoryAwareTrait;
-
-  /**
-   * The team for this report.
-   *
-   * @var \Drupal\apigee_edge_teams\Entity\TeamInterface
-   */
-  protected $team;
+class TeamPrepaidBalanceController extends PrepaidBalanceControllerBase {
 
   /**
    * View prepaid balance and account statements for teams.
@@ -53,117 +40,19 @@ class TeamPrepaidBalanceController extends PrepaidBalanceController {
    *
    * @throws \Exception
    */
-  public function render(TeamInterface $team) {
-    // Set the team for this page call.
-    $this->team = $team;
+  public function teamBalancePage(TeamInterface $team) {
+    // Set the entity for this page call.
+    $this->entity = $team;
 
-    $build = [
-      '#type' => 'container',
-      '#attributes' => [
-        'class' => ['apigee-m10n-prepaid-balance-wrapper'],
-      ],
-      '#attached' => [
-        'library' => [
-          'apigee_m10n/prepaid_balance',
-        ],
-      ],
-    ];
-
-    $build['table'] = [
-      '#type' => 'table',
-      '#header' => $this->buildHeader(),
-      '#title' => $this->getTitle(),
-      '#rows' => [],
-      '#empty' => $this->t('There are no balances available for this @label.', ['@label' => strtolower($team->getEntityType()->getLabel())]),
-      '#cache' => [
-        'contexts' => ['url.path'],
-        'tags' => $this->getCacheTags($team),
-        'max-age' => $max_age = $this->getCacheMaxAge(),
-        'keys' => [static::getCacheId($team, 'prepaid_balances')],
-      ],
-    ];
-    foreach ($this->load() as $currency) {
-      if ($row = $this->buildRow($currency)) {
-        $build['table']['#rows'][$currency->id()] = $row;
-      }
-    }
-
-    // Add a refresh cache form.
-    // TODO: Handle access control for teams.
-    if (TRUE) {
-      $build['refresh_form'] = $this->formBuilder()->getForm(PrepaidBalanceRefreshForm::class, $this->getCacheTags($team));
-    }
-
-    // TODO: Add report download form for teams.
-
-    return $build;
-  }
-
-  /**
-   * Gets the title of the page.
-   */
-  protected function getTitle() {
-    return $this->t('Current prepaid balance');
+    return $this->render();
   }
 
   /**
    * {@inheritdoc}
-   */
-  public function buildHeader() {
-    return [
-      'currency' => $this->t('Account Currency'),
-      'previous_balance' => $this->t('Previous Balance'),
-      'credit' => $this->t('Credit'),
-      'usage' => $this->t('Usage'),
-      'tax' => $this->t('Tax'),
-      'current_balance' => $this->t('Current Balance'),
-    ];
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function buildRow(PrepaidBalanceInterface $balance) {
-    $currency_code = $balance->getCurrency()->getName();
-    return [
-      'currency' => $currency_code,
-      'previous_balance' => $this->formatCurrency($balance->getPreviousBalance(), $currency_code),
-      'credit' => $this->formatCurrency($balance->getTopUps(), $currency_code),
-      'usage' => $this->formatCurrency($balance->getUsage(), $currency_code),
-      'tax' => $this->formatCurrency($balance->getTax(), $currency_code),
-      'current_balance' => $this->formatCurrency($balance->getCurrentBalance(), $currency_code),
-    ];
-  }
-
-  /**
-   * Format an amount using the `monetization` service.
-   *
-   * See: \Drupal\apigee_m10n\MonetizationInterface::formatCurrency().
-   *
-   * @param string $amount
-   *   The money amount.
-   * @param string $currency_code
-   *   Currency code.
-   *
-   * @return string
-   *   The formatted amount as a string.
-   */
-  protected function formatCurrency($amount, $currency_code) {
-    return $this->monetization->formatCurrency($amount, $currency_code);
-  }
-
-  /**
-   * Loads the balances for the listing.
-   *
-   * @return \Apigee\Edge\Api\Monetization\Entity\PrepaidBalanceInterface[]|array
-   *   A list of apigee monetization prepaid balance entities.
-   *
-   * @throws \Exception
    */
   public function load() {
-    $balance_controller = $this->teamControllerFactory()->teamBalanceController($this->team->id());
+    $balance_controller = $this->teamControllerFactory()->teamBalanceController($this->entity->id());
     return $balance_controller->getPrepaidBalance(new \DateTimeImmutable('now'));
   }
 
 }
-
